+++ covasim/population.py
@@ -332,7 +332,7 @@ def make_synthpop(sim, generate=True, layer_mapping=None, **kwargs):
 
     # Handle other input arguments
     pop_size = sim['pop_size']
-    population = sp.make_population(n=pop_size, generate=generate, **kwargs)
+    population = sp.make_population(n=pop_size, generate=generate, rand_seed=sim['rand_seed'], **kwargs)
     uids, ages, sexes, contacts = [], [], [], []
     for uid,person in population.items():
         uids.append(uid)
@@ -362,7 +362,7 @@ def make_synthpop(sim, generate=True, layer_mapping=None, **kwargs):
 
     # Add community contacts
     c_contacts, _ = make_random_contacts(pop_size, {'c':sim['contacts']['c']})
-    for i in range(pop_size):
+    for i in range(int(pop_size)):
         contacts[i]['c'] = c_contacts[i]['c'] # Copy over community contacts -- present for everyone
 
     # Finalize