+++ covasim/utils.py
@@ -6,8 +6,9 @@ Numerical utilities for running Covasim
 
 import numba  as nb # For faster computations
 import numpy  as np # For numerics
-import scipy.stats as sps
-from . import defaults as cvd
+import random # Used only for resetting the seed
+import scipy.stats as sps # For distributions
+from . import defaults as cvd # To set default types
 
 
 # What functions are externally visible -- note, this gets populated in each section below
@@ -189,7 +190,14 @@ def get_pdf(dist=None, par1=None, par2=None):
 
 
 def set_seed(seed=None):
-    ''' Reset the random seed -- complicated because of Numba '''
+    '''
+    Reset the random seed -- complicated because of Numba, which requires special
+    syntax to reset the seed. This function also resets Python's built-in random
+    number generated.
+
+    Args:
+        seed (int): the random seed
+    '''
 
     @nb.njit((nbint,), cache=True)
     def set_seed_numba(seed):
@@ -206,6 +214,7 @@ def set_seed(seed=None):
     if seed is None: # Numba can't accept a None seed, so use our just-reinitialized Numpy stream to generate one
         seed = np.random.randint(1e9)
     set_seed_numba(seed)
+    random.seed(seed) # Finally, reset Python's built-in random number generator, just in case (used by SynthPops)
 
     return